Completed suggested changes
