+++ src/app/beer_garden/api/entry_point.py
@@ -52,7 +52,9 @@ class EntryPoint(object):
         module = getattr(beer_garden.api, module_name)
         return EntryPoint(module_name, module.run, signal_handler=module.signal_handler)
 
-    def start(self, context: BaseContext, log_queue: Queue, events_queue: Queue) -> None:
+    def start(
+        self, context: BaseContext, log_queue: Queue, events_queue: Queue
+    ) -> None:
         """Start the entry point process
 
         Args:
@@ -136,7 +138,8 @@ class EntryPoint(object):
         beer_garden.config.assign(config)
 
         # Then setup Event Manager Queue
-        beer_garden.establish_events_queue(event_queue)
+        # beer_garden.establish_events_queue(event_queue)
+        beer_garden.events.events_manager.establish_events_queue(event_queue)
 
         # Then set up logging to push everything back to the main process
         beer_garden.log.setup_entry_point_logging(log_queue)
