+++ src/app/beer_garden/systems.py
@@ -11,7 +11,7 @@ from time import sleep
 import beer_garden
 import beer_garden.db.api as db
 import beer_garden.queue.api as queue
-from beer_garden.events import publish_event
+from beer_garden.events.events_manager import publish_event
 from beer_garden.queue.rabbit import get_routing_key
 
 REQUEST_FIELDS = set(SystemSchema.get_attribute_names())
